+++ src/main/java/edu/cornell/kfs/paymentworks/service/PaymentWorksAchService.java
@@ -1,10 +1,12 @@
 package edu.cornell.kfs.paymentworks.service;
 
+import java.util.List;
+
 import org.kuali.kfs.pdp.businessobject.PayeeACHAccount;
 
 import edu.cornell.kfs.paymentworks.xmlObjects.PaymentWorksVendorUpdatesDTO;
 
-public interface PaymentWorksAchConversionService {
+public interface PaymentWorksAchService {
 	
 	/**
 	 * Creates a Payee ACH Account object based on a DTO pulled in from PaymentWorks
@@ -22,4 +24,6 @@ public interface PaymentWorksAchConversionService {
 	 * @return
 	 */
 	PayeeACHAccount createPayeeAchAccount(PayeeACHAccount payeeAchAccountOld, String routingNumber, String accountNumber);
+	
+	boolean processACHUpdates(List<PaymentWorksVendorUpdatesDTO> achUpdates);
 }
